iabank/backend/pyproject.toml
[tool.poetry]
name = "iabank"
version = "0.1.0"
description = "Sistema de gestão de empréstimos moderno e eficiente"
authors = ["IABANK Team <team@iabank.com>"]
readme = "README.md"
packages = [{include = "iabank", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.0"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-cors-headers = "^4.0.0"
django-extensions = "^3.2.0"
psycopg2-binary = "^2.9.6"
redis = "^4.5.0"
celery = "^5.2.0"
pydantic = "^2.0.0"
djangorestframework-simplejwt = "^5.2.0"
factory-boy = "^3.2.1"
structlog = "^23.1.0"
django-prometheus = "^2.3.1"
django-filter = "^23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
pytest-django = "^4.5.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
ruff = "^0.0.270"
pre-commit = "^3.3.0"
mypy = "^1.3.0"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"
pytest-mock = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "iabank.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["src"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing --verbose --cov-config=.coveragerc"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "iabank.settings"